YanaBot 3.0 update

+++ tg_bot/modules/welcome.py
@@ -17,6 +17,9 @@ from tg_bot.modules.helper_funcs.string_handling import markdown_parser, \
     escape_invalid_curly_brackets
 from tg_bot.modules.log_channel import loggable
 
+from tg_bot.modules.feds import welcome_fed
+
+
 VALID_WELCOME_FORMATTERS = ['first', 'last', 'fullname', 'username', 'id', 'count', 'chatname', 'mention']
 
 ENUM_FUNC_MAP = {
@@ -131,6 +134,10 @@ def new_member(bot: Bot, update: Update):
                 sent = send(update, res, keyboard,
                             sql.DEFAULT_WELCOME.format(first=first_name))  # type: Optional[Message]
 
+                #feds
+                if welcome_fed(bot, update) == True:
+                    continue
+
                 #Clean service welcome
                 if sql.clean_service(chat.id) == True:
                     bot.delete_message(chat.id, update.message.message_id)
@@ -145,7 +152,7 @@ def new_member(bot: Bot, update: Update):
 
                 #Add "I'm not bot button if enabled hard security"
                 if sql.welcome_security(chat.id) == "hard":
-                    update.effective_message.reply_text("Hi {}, click on button below for been unmuted.".format(new_mem.first_name), 
+                    update.effective_message.reply_text("Hi {}, click on button below to prove you're human.".format(new_mem.first_name), 
                          reply_markup=InlineKeyboardMarkup([[InlineKeyboardButton(text="I'm not a bot!", 
                          callback_data="check_bot_({})".format(new_mem.id)) ]]))
                     #Mute user
@@ -182,7 +189,7 @@ def check_bot_button(bot: Bot, update: Update):
         bot.deleteMessage(chat.id, message.message_id)
     else:
         print("NO")
-        query.answer(text="You not a new user!")
+        query.answer(text="You're not a new user!")
     #TODO need kick users after 2 hours and remove message 
 
 @run_async
@@ -453,10 +460,10 @@ def security(bot: Bot, update: Update, args: List[str]) -> str:
             update.effective_message.reply_text("Disabled welcome security")
         elif(var == "soft"):
             sql.set_welcome_security(chat.id, "soft")
-            update.effective_message.reply_text("I will restrict user to send media for 24 hours")
+            update.effective_message.reply_text("I will restrict user's permission to send media for 24 hours")
         elif(var == "hard"):
             sql.set_welcome_security(chat.id, "hard")
-            update.effective_message.reply_text("I will mute user when he don't click on button")
+            update.effective_message.reply_text("New users will be muted if they do not click on the button")
         else:
             update.effective_message.reply_text("Please enter `off`/`no`/`soft`/`hard`!", parse_mode=ParseMode.MARKDOWN)
     else:
@@ -474,7 +481,7 @@ def cleanservice(bot: Bot, update: Update, args: List[str]) -> str:
             print(var)
             if (var == "no" or var == "off"):
                 sql.set_clean_service(chat.id, False)
-                update.effective_message.reply_text("I leave service messages")
+                update.effective_message.reply_text("I'll leave service messages")
             elif(var == "yes" or var == "on"):
                 sql.set_clean_service(chat.id, True)
                 update.effective_message.reply_text("I will clean service messages")
@@ -511,39 +518,43 @@ def __chat_settings__(bot, update, chat, chatP, user):
 
 
 __help__ = """
-Your group's welcome/goodbye messages can be personalised in multiple ways. If you want the messages \
-to be individually generated, like the default welcome message is, you can use *these* variables:
- - `{{first}}`: this represents the user's *first* name
- - `{{last}}`: this represents the user's *last* name. Defaults to *first name* if user has no last name.
- - `{{fullname}}`: this represents the user's *full* name. Defaults to *first name* if user has no last name.
- - `{{username}}`: this represents the user's *username*. Defaults to a *mention* of the user's first name if has no username.
- - `{{mention}}`: this simply *mentions* a user - tagging them with their first name.
- - `{{id}}`: this represents the user's *id*.
- - `{{count}}`: this represents the user's *member number*.
- - `{{chatname}}`: this represents the *current chat name*.
-Each variable MUST be surrounded by `{{}}` to be replaced.
-Welcome messages also support markdown, so you can make any elements bold/italic/code/links. \
-Buttons are also supported, so you can make your welcomes look awesome with some nice intro \
-buttons. To create a button linking to your rules, use this: `[Rules](buttonurl://t.me/{}?start=group_id)`. \
-Simply replace `group_id` with your group's id, which can be obtained via /id, and you're good to \
-go. Note that group ids are usually preceded by a `-` sign; this is required, so please don't \
-remove it. \
-If you're feeling fun, you can even set images/gifs/videos/voice messages as the welcome message by \
-replying to the desired media, and calling /setwelcome.
+Give your members a warm welcome with the greetings module! Or a sad goodbye... Depends!
 
-*Admin only:*
- - /welcome <on/off>: enable/disable welcome messages.
- - /welcome: shows current welcome settings.
- - /welcome noformat: shows current welcome settings, without the formatting - useful to recycle your welcome messages!
- - /goodbye -> same usage and args as /welcome.
- - /setwelcome <sometext>: set a custom welcome message. If used replying to media, uses that media.
- - /setgoodbye <sometext>: set a custom goodbye message. If used replying to media, uses that media.
- - /resetwelcome: reset to the default welcome message.
- - /resetgoodbye: reset to the default goodbye message.
- - /cleanwelcome <on/off>: On new member, try to delete the previous welcome message to avoid spamming the chat.
+Available commands are:
+ - /welcome <on/off/yes/no>: enables/disables welcome messages. If no option is given, returns the current welcome message and welcome settings. 
+ - /goodbye <on/off/yes/no>: enables/disables goodbye messages. If no option is given, returns  the current goodbye message and goodbye settings.
+ - /setwelcome <message>: sets your new welcome message! Markdown and buttons are supported, as well as fillings.
+ - /resetwelcome: resets your welcome message to default; deleting any changes you've made.
+ - /setgoodbye <message>: sets your new goodbye message! Markdown and buttons are supported, as well as fillings.
+ - /resetgoodbye: resets your goodbye message to default; deleting any changes you've made.
+ - /cleanwelcome <on/off/yes/no>: deletes old welcome messages; when a new person joins, the old message is deleted.
  - /cleanservice <on/off/yes/no>: deletes all service message; those are the annoying "x joined the group" you see when people join.
- - /welcomesecurity <off/soft/hard>: soft - restrict user send media files for 24 hours, hard - restict user send messages while him don't click on button \"I'm not bot\"
-""".format(dispatcher.bot.username)
+ - /welcomesecurity <off/soft/hard>: soft - restrict user's permission to send media files for 24 hours, hard - restict user's permission to send messages until they click on the button \"I'm not a bot\"
+
+
+Fillings:
+As mentioned, you can use certain tags to fill in your welcome message with user or chat info; there are:
+{first}: The user's first name.
+{last}: The user's last name.
+{fullname}: The user's full name.
+{username}: The user's username; if none is available, mentions the user.
+{mention}: Mentions the user, using their firstname.
+{id}: The user's id.
+{chatname}: The chat's name.
+
+An example of how to use fillings would be to set your welcome, via:
+/setwelcome Hey there {first}! Welcome to {chatname}.
+
+You can enable/disable welcome messages as such:
+/welcome off
+
+If you want to save an image, gif, or sticker, or any other data, do the following:
+/setwelcome while replying to a sticker or whatever data you'd like. This data will now be sent to welcome new users.
+
+Tip: use /welcome noformat to retrieve the unformatted welcome message.
+This will retrieve the welcome message and send it without formatting it; getting you the raw markdown, allowing you to make easy edits.
+This also works with /goodbye.
+"""
 
 
 __mod_name__ = "Welcomes/Goodbyes"
